同步locally.json 编辑结果

+++ utils/FileUtils.py
@@ -3,16 +3,13 @@ from pathlib import Path
 
 
 def createNewFile(path):
-    print("createNewFile...")
-    print(path)
     file = Path(path)
     file.touch()
 
 
 def deleteFile(path):
-    print("deleteFile...")
-    print(path)
-    os.remove(path)
+    if os.path.exists(path):
+        os.remove(path)
 
 
 def writeString2File(text: str, path: str):
